ajout de l'event click pour affichage des spécificités
