+++ src/App.jsx
@@ -1,8 +1,6 @@
 import React from 'react';
-// import logo from './logo.svg';
-// import Alien from './component/alien.js';
 import './App.css';
-import Alienlist from './component/alienlist.js';
+import Alienlist from './Components/Alienlist.jsx';
 
 function App() {
   return (
@@ -13,25 +11,3 @@ function App() {
 }
 
 export default App;
-
-// function App() {
-//   return (
-//     <div className="App">
-//       <header className="App-header">
-//         <img src={logo} className="App-logo" alt="logo" />
-//         <p>Edit
-//           <code>src/App.js</code>
-//           and save to reload.
-//           </p>
-//         <a
-//           className="App-link"
-//           href="https://reactjs.org"
-//           target="_blank"
-//           rel="noopener noreferrer"
-//         >
-//           Learn React
-//         </a>
-//       </header>
-//     </div>
-//   );
-// }
