+++ src/Components/Alienlist.jsx
@@ -1,23 +1,29 @@
 import React from 'react';
-import Alien from './Alien.js';
+import Alien from './Alien.jsx';
 
 const alienuser = [
   {
     image: 'https://rickandmortyapi.com/api/character/avatar/15.jpeg',
     name: 'Alien Rick',
-    id: 15
+    id: 15,
+    gender: 'Male',
+    species: 'Alien'
   },
 
   {
     image: 'https://rickandmortyapi.com/api/character/avatar/14.jpeg',
     name: 'Alien Morty',
-    id: 14
+    id: 14,
+    gender: 'Male',
+    species: 'Alien'
   },
 
   {
     image: 'https://rickandmortyapi.com/api/character/avatar/13.jpeg',
     name: 'Alien Googah',
-    id: 13
+    id: 13,
+    gender: 'Male',
+    species: 'Alien'
   }
 ];
 
@@ -25,7 +31,7 @@ function Alienlist(props) {
   return (
     <ul className="bloc">
       {alienuser.map(item => (
-        <Alien image={item.image} name={item.name} id={item.id} />
+        <Alien alienInfos={item} />
       ))}
     </ul>
   );
