--- src/component/alien.css
@@ -1,19 +1,0 @@
-@media (max-width: 720px) {
-  .image {
-    border-radius: 15%;
-    height: auto;
-    padding: auto;
-  }
-
-  .alien {
-    display: flex;
-    flex-direction: column;
-    text-align: center;
-    font-size: 50%;
-  }
-
-  .name {
-    width: 100%;
-    height: auto;
-  }
-}
