--- src/component/alienlist.css
@@ -1,3 +1,0 @@
-.bloc {
-  padding: auto;
-}
